Refactor buttonCss
