+++ src/containers/MyNdla/FavoriteSubjects/FavoriteSubjectsPage.tsx
@@ -21,7 +21,8 @@ import SubjectLink from "../../AllSubjectsPage/SubjectLink";
 import MyNdlaPageWrapper from "../components/MyNdlaPageWrapper";
 import MyNdlaTitle from "../components/MyNdlaTitle";
 import SettingsMenu from "../components/SettingsMenu";
-import { buttonCss, iconCss } from "../Folders/FoldersPage";
+import { buttonCss } from "../components/toolbarStyles";
+import { iconCss } from "../Folders/FoldersPage";
 import { useSubjects } from "../subjectQueries";
 
 const StyledSubjectLink = styled(SubjectLink)`
