+++ src/containers/MyNdla/components/MenuModalContent.tsx
@@ -18,10 +18,11 @@ import { SafeLinkButton } from "@ndla/safelink";
 import { Text } from "@ndla/typography";
 import NavigationLink from "./NavigationLink";
 import { BellIcon } from "./NotificationButton";
+import { buttonCss } from "./toolbarStyles";
 import { AuthContext } from "../../../components/AuthenticationContext";
 import { routes } from "../../../routeHelpers";
 import { useTemporaryArenaNotifications } from "../Arena/components/temporaryNodebbHooks";
-import { ViewType, buttonCss } from "../Folders/FoldersPage";
+import { ViewType } from "../Folders/FoldersPage";
 import { OutletContext, menuLinks } from "../MyNdlaLayout";
 
 const MenuItem = styled.li`
